Make it so TeamCity running the tests gets pretty test results
This adds the teamcity-messages packages which extends py.test and flake8 in
order to make it so messages from the tools will be picked up by the CI system
directly.

This also passes TEAMCITY_VERSION through the tox.ini environment variables. The
teamcity-messages automatically enables / starts itself when TEAMCITY_VERSION is
in the environment, which is only set on TeamCity build agents, making for
minimal CI changes being necessary, while leaving it compatible with other CI
system message formatting plugins.